metrics-collector/requirements.txt
fastapi
fastapi_utils
uvicorn
influxdb_client
load_dotenv
requests
pyyaml
